requirements.txt
fastapi==0.91.0
hypercorn[uvloop,httptools]
httpx==0.24.1
numpy==1.23.3
pillow==10.0.0
pydantic==1.10.2
deep_translator==1.8.3
python_weather==0.4.3
pilmoji==2.0.3
pydash==7.0.6
colour==0.1.5